# docs/migration/contract/openapi-legacy-baseline.yaml
openapi: 3.1.0
info:
  title: Gia Pha Legacy Backend HTTP Contract (Frozen Baseline)
  version: 1.0.0-legacy-baseline
  description: |
    Frozen contract of every Next.js App Router `/api/**` handler prior to the
    Spring Boot migration (Task 1 of spring-boot-backend-migration).

    Conventions frozen from the legacy implementation:
    * **Error envelope** — every error response is `{"ok": false, "error": {"code", "message", "details?"}}`.
    * **Mixed success envelopes** — most successful responses return the raw
      resource JSON (no envelope). The exceptions that wrap in
      `{"ok": true, "data": ...}` are: `POST /api/auth/register`,
      `PATCH /api/trees/{treeId}/memberships/{userId}`,
      `POST /api/media/upload`, and share-link management responses.
    * **Timestamps** — ISO-8601 UTC strings with milliseconds (`.toISOString()` form).
    * **Authentication** — NextAuth JWT session cookie
      (`next-auth.session-token` / `__Secure-next-auth.session-token`).
      Unauthenticated requests receive `401 UNAUTHORIZED`.
    * **RBAC** — tree-scoped roles ADMIN/EDITOR/VIEWER, owner is immutable
      effective ADMIN. Missing tree → `404 TREE_NOT_FOUND`, missing
      permission → `403 FORBIDDEN`.
    * **Tree discovery** — detail routes for members/relationships/events/media/
      albums accept optional `?treeId=`; without it the legacy handler scans
      every tree to locate the resource. Compatibility_API must preserve this.
...
servers:
  - url: /
...
paths:
...
  /api/trees/{treeId}:
    parameters: [ { $ref: '#/components/parameters/treeId' } ]
    get:
      tags: [trees]
      operationId: getTree
      description: Requires READ. Returns tree plus members array.
      responses:
        '200':
          description: Tree with members.
          content:
            application/json:
              schema:
                allOf:
                  - { $ref: '#/components/schemas/FamilyTree' }
                  - type: object
                    properties:
                      members: { type: array, items: { $ref: '#/components/schemas/Member' } }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
    patch:
      tags: [trees]
      operationId: updateTree
      description: Requires UPDATE.
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/UpdateTreeInput' } } }
      responses:
        '200': { description: Raw updated tree., content: { application/json: { schema: { $ref: '#/components/schemas/FamilyTree' } } } }
        '400': { $ref: '#/components/responses/ValidationError' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
    delete:
      tags: [trees]
      operationId: deleteTree
      description: Requires DELETE (ADMIN). Removes tree, memberships and all tree-scoped blobs.
      responses:
        '204': { description: Deleted. Empty body. }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
...
  /api/trees/{treeId}/relationships:
    parameters: [ { $ref: '#/components/parameters/treeId' } ]
    get:
      tags: [relationships]
      operationId: listRelationships
      description: Requires READ. `?memberId=` returns RelationshipView list for that member.
      parameters:
        - { name: memberId, in: query, required: false, schema: { type: string } }
      responses:
        '200':
          description: Raw array (Relationship[] or RelationshipView[] when memberId given).
          content: { application/json: { schema: { type: array, items: { $ref: '#/components/schemas/Relationship' } } } }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
    post:
      tags: [relationships]
      operationId: createRelationship
      description: |
        Requires CREATE. Symmetric types (SPOUSE/SIBLING/CUSTOM) are stored with
        canonical endpoint order (source <= target). Duplicates by logical key
        rejected 409 DUPLICATE; PARENT_CHILD cycles rejected 409 CYCLE_DETECTED.
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/CreateRelationshipInput' } } } }
      responses:
        '201': { description: Raw created relationship (normalized)., content: { application/json: { schema: { $ref: '#/components/schemas/Relationship' } } } }
        '400': { $ref: '#/components/responses/ValidationError' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { description: DUPLICATE or CYCLE_DETECTED, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorEnvelope' } } } }
...
  /api/media/upload:
    post:
      tags: [media]
      operationId: uploadMedia
      description: |
        Requires CREATE on target tree. multipart/form-data with fields:
        file (required), treeId (required), memberId(s), eventId(s), albumId,
        caption, takenAt, isAvatar. JPEG/PNG/WebP/PDF, max 10 MiB, magic-byte
        validated; image uploads produce a 480px WebP q78 thumbnail. Vietnamese
        error messages are part of the contract.
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file, treeId]
              properties:
                file: { type: string, format: binary }
                treeId: { type: string }
                memberId: { type: string }
                memberIds: { type: string, description: JSON array or CSV }
                eventId: { type: string }
                eventIds: { type: string }
                albumId: { type: string }
                caption: { type: string, maxLength: 2000 }
                takenAt: { type: string, format: date-time }
                isAvatar: { type: string, enum: ['true', 'false'] }
      responses:
        '201':
          description: Wrapped media metadata.
          content:
            application/json:
              schema:
                type: object
                required: [ok, data]
                properties:
                  ok: { const: true }
                  data: { $ref: '#/components/schemas/MediaMetadata' }
        '400': { $ref: '#/components/responses/ValidationError' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '413': { description: FILE_TOO_LARGE, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorEnvelope' } } } }
        '415': { description: INVALID_FILE_TYPE, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorEnvelope' } } } }
  /api/trees/{treeId}/media:
    parameters: [ { $ref: '#/components/parameters/treeId' } ]
    get:
      tags: [media]
      operationId: listMedia
      description: |
        Requires READ. At most one of memberId/eventId/albumId; more than one
        → 400 with Vietnamese message. Sorted by takenAt||uploadedAt desc.
      parameters:
        - { name: memberId, in: query, required: false, schema: { type: string } }
        - { name: eventId, in: query, required: false, schema: { type: string } }
        - { name: albumId, in: query, required: false, schema: { type: string } }
      responses:
        '200': { description: Raw array., content: { application/json: { schema: { type: array, items: { $ref: '#/components/schemas/MediaMetadata' } } } } }
        '400': { $ref: '#/components/responses/ValidationError' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
  /api/media/{mediaId}:
    parameters:
      - { name: mediaId, in: path, required: true, schema: { type: string } }
      - { name: treeId, in: query, required: false, schema: { type: string } }
    get:
      tags: [media]
      operationId: getMedia
      responses:
        '200': { description: Raw metadata., content: { application/json: { schema: { $ref: '#/components/schemas/MediaMetadata' } } } }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
    delete:
      tags: [media]
      operationId: deleteMedia
      description: Metadata removed first; blob cleanup failures logged, not surfaced.
      responses:
        '204': { description: Deleted. Empty body. }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
...
  /api/trees/{treeId}/albums:
    parameters: [ { $ref: '#/components/parameters/treeId' } ]
    get:
      tags: [albums]
      operationId: listAlbums
      description: Requires READ. Sorted by createdAt desc.
      responses:
        '200': { description: Raw array., content: { application/json: { schema: { type: array, items: { $ref: '#/components/schemas/Album' } } } } }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
    post:
      tags: [albums]
      operationId: createAlbum
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/CreateAlbumInput' } } } }
      responses:
        '201': { description: Raw created album., content: { application/json: { schema: { $ref: '#/components/schemas/Album' } } } }
        '400': { $ref: '#/components/responses/ValidationError' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
  /api/albums/{albumId}:
    parameters:
      - { name: albumId, in: path, required: true, schema: { type: string } }
      - { name: treeId, in: query, required: false, schema: { type: string } }
    patch:
      tags: [albums]
      operationId: updateAlbum
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/UpdateAlbumInput' } } } }
      responses:
        '200': { description: Raw updated album., content: { application/json: { schema: { $ref: '#/components/schemas/Album' } } } }
        '400': { $ref: '#/components/responses/ValidationError' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
    delete:
      tags: [albums]
      operationId: deleteAlbum
      description: Album removed; linked media keep existing but albumId is cleared.
      responses:
        '204': { description: Deleted. Empty body. }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }

  # ---------------------------------------------------------------- search
...
components:
  securitySchemes:
    nextAuthSession:
      type: apiKey
      in: cookie
      name: next-auth.session-token
  parameters:
    treeId:
      name: treeId
      in: path
      required: true
      schema: { type: string }
  responses:
    Unauthorized:
      description: Missing/invalid session (code UNAUTHORIZED).
      content: { application/json: { schema: { $ref: '#/components/schemas/ErrorEnvelope' } } }
    Forbidden:
      description: Authenticated but lacking the tree permission (code FORBIDDEN).
      content: { application/json: { schema: { $ref: '#/components/schemas/ErrorEnvelope' } } }
    NotFound:
      description: Resource or tree not found (code NOT_FOUND / TREE_NOT_FOUND).
      content: { application/json: { schema: { $ref: '#/components/schemas/ErrorEnvelope' } } }
    ValidationError:
      description: Invalid input (code VALIDATION_ERROR / INVALID_INPUT / INVALID_FILTER; details carries Zod flatten()).
      content: { application/json: { schema: { $ref: '#/components/schemas/ErrorEnvelope' } } }
...
  schemas:
    ErrorEnvelope:
      type: object
      required: [ok, error]
      properties:
        ok: { const: false }
        error:
          type: object
          required: [code, message]
          properties:
            code: { type: string }
            message: { type: string }
            details: {}
    TreeRole: { type: string, enum: [ADMIN, EDITOR, VIEWER] }
    Gender: { type: string, enum: [MALE, FEMALE, OTHER] }
    RelationshipType: { type: string, enum: [PARENT_CHILD, SPOUSE, SIBLING, ADOPTED, CUSTOM] }
...
    FamilyTree:
      type: object
      required: [id, name, ownerId, memberships, createdAt, updatedAt]
      properties:
        id: { type: string }
        name: { type: string, minLength: 1, maxLength: 200 }
        description: { type: string, maxLength: 2000 }
        ownerId: { type: string }
        memberships: { type: array, items: { $ref: '#/components/schemas/TreeMembership' } }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    TreeMembership:
      type: object
      required: [userId, treeId, role, joinedAt]
      properties:
        userId: { type: string }
        treeId: { type: string }
        role: { $ref: '#/components/schemas/TreeRole' }
        joinedAt: { type: string, format: date-time }
...
    UpdateTreeInput:
      type: object
      properties:
        name: { type: string, minLength: 1, maxLength: 200 }
        description: { type: string, maxLength: 2000 }
    Member:
      type: object
      required: [id, treeId, firstName, lastName, fullName, gender, isAlive, createdAt, updatedAt]
      properties:
        id: { type: string }
        treeId: { type: string }
        firstName: { type: string, minLength: 1, maxLength: 100 }
        lastName: { type: string, minLength: 1, maxLength: 100 }
        fullName: { type: string, minLength: 1, maxLength: 200 }
        nickname: { type: string, maxLength: 100 }
        gender: { $ref: '#/components/schemas/Gender' }
        dateOfBirth: { type: string, description: ISO date or date-time }
        dateOfDeath: { type: string }
        isAlive: { type: boolean }
        placeOfBirth: { type: string, maxLength: 500 }
        currentAddress: { type: string, maxLength: 500 }
        occupation: { type: string, maxLength: 200 }
        education: { type: string, maxLength: 200 }
        phone: { type: string, maxLength: 20 }
        email: { type: string, maxLength: 254 }
        biography: { type: string, maxLength: 5000 }
        avatarUrl: { type: string, deprecated: true }
        avatarMediaId: { type: string, description: Must reference same-tree image media }
        generation: { type: integer }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
...
    Relationship:
      type: object
      required: [id, treeId, sourceMemberId, targetMemberId, type, createdAt, updatedAt]
      properties:
        id: { type: string }
        treeId: { type: string }
        sourceMemberId: { type: string }
        targetMemberId: { type: string }
        type: { $ref: '#/components/schemas/RelationshipType' }
        customType: { type: string, maxLength: 100, description: Required when type=CUSTOM }
        startDate: { type: string }
        endDate: { type: string }
        notes: { type: string, maxLength: 2000 }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    CreateRelationshipInput:
      type: object
      required: [sourceMemberId, targetMemberId, type]
      description: source != target; CUSTOM requires customType.
...
    MediaMetadata:
      type: object
      required: [id, treeId, memberIds, eventIds, filename, originalName, mimeType, fileSize, blobUrl, uploadedAt]
      properties:
        id: { type: string }
        treeId: { type: string }
        memberId: { type: string, deprecated: true }
        eventId: { type: string, deprecated: true }
        memberIds: { type: array, items: { type: string } }
        eventIds: { type: array, items: { type: string } }
        albumId: { type: string }
        filename: { type: string }
        originalName: { type: string, maxLength: 255 }
        mimeType: { type: string, enum: [image/jpeg, image/png, image/webp, application/pdf] }
        fileSize: { type: integer, maximum: 10485760 }
        blobUrl: { type: string, description: Private storage URL — never usable directly by clients }
        thumbnailUrl: { type: string }
        contentUrl: { type: string, description: '/api/media/{id}/content?treeId=...' }
        thumbnailContentUrl: { type: string }
        caption: { type: string, maxLength: 2000 }
        takenAt: { type: string, format: date-time }
        uploadedAt: { type: string, format: date-time }
    Album:
      type: object
      required: [id, treeId, name, createdAt, updatedAt]
      properties:
        id: { type: string }
        treeId: { type: string }
        name: { type: string, minLength: 1, maxLength: 200 }
        description: { type: string, maxLength: 2000 }
        coverMediaId: { type: string }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    CreateAlbumInput: { type: object, required: [name] }
    UpdateAlbumInput: { type: object }
...
security:
  - nextAuthSession: []